backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
flask-jwt-extended==4.5.3
werkzeug==2.3.7
python-dotenv==1.0.0
pillow==10.0.1
psycopg2-binary==2.9.7
bcrypt==4.0.1
mutagen==1.47.0.23
psycopg2-binary==2.9.9
redis==5.0.1
requests==2.31.0
python-multipart==0.0.6
passlib[bcrypt]==1.7.4
asyncpg==0.29.0
click==8.1.7
PyYAML==6.0.1
pydantic==2.5.0
pydantic-settings==2.1.0